# docs/swagger.yaml
basePath: /
definitions:
  giiku-camp_internal_controller_render.Error:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: integer
    type: object
  giiku-camp_internal_usecase_request.RefreshTokenReq:
    properties:
      refresh_token:
        type: string
    type: object
  giiku-camp_internal_usecase_request.SignInReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  giiku-camp_internal_usecase_request.SignUpReq:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  giiku-camp_internal_usecase_response.RefreshTokenRes:
    properties:
      token:
        properties:
          access_token:
            type: string
          refresh_token:
            type: string
        type: object
    type: object
  giiku-camp_internal_usecase_response.SignInRes:
    properties:
      token:
        properties:
          access_token:
            type: string
          refresh_token:
            type: string
        type: object
      user:
        $ref: '#/definitions/giiku-camp_internal_usecase_response.UserRes'
    type: object
  giiku-camp_internal_usecase_response.SignUpRes:
    properties:
      token:
        properties:
          access_token:
            type: string
          refresh_token:
            type: string
        type: object
      user:
        $ref: '#/definitions/giiku-camp_internal_usecase_response.UserRes'
    type: object
  giiku-camp_internal_usecase_response.UserRes:
    properties:
      created_at:
        type: string
      email:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the Giiku Camp API server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Giiku Camp API
  version: "1.0"
paths:
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: トークンの更新
      parameters:
      - description: Refresh token
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/giiku-camp_internal_usecase_request.RefreshTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed
          schema:
            $ref: '#/definitions/giiku-camp_internal_usecase_response.RefreshTokenRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/giiku-camp_internal_controller_render.Error'
      summary: トークンの更新
      tags:
      - User
  /auth/signin:
    post:
      consumes:
      - application/json
      description: ユーザー認証
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/giiku-camp_internal_usecase_request.SignInReq'
      produces:
      - application/json
      responses:
        "200":
          description: User authenticated
          schema:
            $ref: '#/definitions/giiku-camp_internal_usecase_response.SignInRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/giiku-camp_internal_controller_render.Error'
      summary: ユーザー認証
      tags:
      - User
  /auth/signup:
    post:
      consumes:
      - application/json
      description: ユーザー登録
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/giiku-camp_internal_usecase_request.SignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: User created
          schema:
            $ref: '#/definitions/giiku-camp_internal_usecase_response.SignUpRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/giiku-camp_internal_controller_render.Error'
      summary: ユーザー登録
      tags:
      - User
  /ping:
    get:
      description: Returns "pong" if server is working
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Ping the server
      tags:
      - Health
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"